requirements.txt
Flask
pdfminer.six
pytesseract
Pillow